redispy-entraid/requirements.txt
PyJWT~=2.9.0
pytest
msal~=1.31.0